refactor: 자동차 이름 VO 생성
